Fix cluster query memory leak (#4335)
